+++ src/ts/AdUnits/MRAIDAdUnit.ts
@@ -9,7 +9,6 @@ import { MRAID, IOrientationProperties } from 'Views/MRAID';
 import { AdUnitContainer, ForceOrientation } from 'AdUnits/Containers/AdUnitContainer';
 import { Platform } from 'Constants/Platform';
 import { HTML } from 'Models/Assets/HTML';
-import { Diagnostics } from 'Utilities/Diagnostics';
 
 export class MRAIDAdUnit extends AbstractAdUnit {
 
@@ -20,6 +19,8 @@ export class MRAIDAdUnit extends AbstractAdUnit {
 
     private _onShowObserver: IObserver0;
     private _onSystemKillObserver: IObserver0;
+    private _onSystemInterruptObserver: any;
+    private _onPauseObserver: any;
     private _additionalTrackingEvents: { [eventName: string]: string[] };
 
     constructor(nativeBridge: NativeBridge, container: AdUnitContainer, sessionManager: SessionManager, placement: Placement, campaign: MRAIDCampaign, mraid: MRAID, options: any) {
@@ -59,6 +60,8 @@ export class MRAIDAdUnit extends AbstractAdUnit {
 
         this._onShowObserver = this._container.onShow.subscribe(() => this.onShow());
         this._onSystemKillObserver = this._container.onSystemKill.subscribe(() => this.onSystemKill());
+        this._onSystemInterruptObserver = this._container.onSystemInterrupt.subscribe((interruptStarted) => this.onSystemInterrupt(interruptStarted));
+        this._onPauseObserver = this._container.onAndroidPause.subscribe(() => this.onSystemPause());
 
         return this._container.open(this, false, this._orientationProperties.allowOrientationChange, this._orientationProperties.forceOrientation, true, this._options);
     }
@@ -71,6 +74,8 @@ export class MRAIDAdUnit extends AbstractAdUnit {
 
         this._container.onShow.unsubscribe(this._onShowObserver);
         this._container.onSystemKill.unsubscribe(this._onSystemKillObserver);
+        this._container.onSystemInterrupt.unsubscribe(this._onSystemInterruptObserver);
+        this._container.onAndroidPause.unsubscribe(this._onPauseObserver);
 
         this._mraid.hide();
 
@@ -111,20 +116,9 @@ export class MRAIDAdUnit extends AbstractAdUnit {
         this.sendTrackingEvent('click');
     }
 
-    public sendClickAttribution(): void {
-        const clickAttributionUrl = (<MRAIDCampaign>this.getCampaign()).getClickAttributionUrl();
-        if(clickAttributionUrl) {
-            this._sessionManager.getEventManager().clickAttributionEvent(clickAttributionUrl, false);
-            if((<MRAIDCampaign>this.getCampaign()).getClickAttributionUrlFollowsRedirects()) {
-                Diagnostics.trigger('mraid_click_attribution_url_follows_redirects', {
-                    url: clickAttributionUrl,
-                    followsRedirects: true
-                });
-            }
-        }
-    }
-
     private onShow() {
+        this._mraid.setViewableState(true);
+
         if(AbstractAdUnit.getAutoClose()) {
             setTimeout(() => {
                 this.setFinishState(FinishState.COMPLETED);
@@ -137,6 +131,22 @@ export class MRAIDAdUnit extends AbstractAdUnit {
         if(this.isShowing()) {
             this.setFinishState(FinishState.SKIPPED);
             this.hide();
+        }
+    }
+
+    private onSystemInterrupt(interruptStarted: boolean): void {
+        if(this.isShowing()) {
+            if(interruptStarted) {
+                this._mraid.setViewableState(false);
+            } else {
+                this._mraid.setViewableState(true);
+            }
+        }
+    }
+
+    private onSystemPause(): void {
+        if(this.isShowing()) {
+            this._mraid.setViewableState(false);
         }
     }