+++ src/ts/AdUnits/VideoAdUnit.ts
@@ -31,15 +31,19 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
     private _deviceInfo: DeviceInfo;
     private _videoOrientation: 'landscape' | 'portrait' | undefined;
     private _lowMemory: boolean;
+    private _prepareCalled: boolean;
+    private _videoReady: boolean;
 
     constructor(nativeBridge: NativeBridge, forceOrientation: ForceOrientation, container: AdUnitContainer, placement: Placement, campaign: Campaign, video: Video, overlay: Overlay, deviceInfo: DeviceInfo, options: any) {
         super(nativeBridge, forceOrientation, container, placement, campaign);
 
         this._video = video;
+        this._videoReady = false;
         this._active = false;
         this._overlay = overlay;
         this._deviceInfo = deviceInfo;
         this._options = options;
+        this._prepareCalled = false;
 
         this._lowMemory = false;
     }
@@ -71,6 +75,22 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
         return this._container.close().then(() => {
             this.onClose.trigger();
         });
+    }
+
+    public isVideoReady(): boolean {
+        return this._videoReady;
+    }
+
+    public setVideoReady(ready: boolean): void {
+        this._videoReady = ready;
+    }
+
+    public isPrepareCalled(): boolean {
+        return this._prepareCalled;
+    }
+
+    public setPrepareCalled(prepareCalled: boolean): void {
+        this._prepareCalled = prepareCalled;
     }
 
     public isCached(): boolean {
@@ -120,6 +140,7 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
             }
 
             this.getValidVideoUrl().then(url => {
+                this.setPrepareCalled(true);
                 this._nativeBridge.VideoPlayer.prepare(url, new Double(this._placement.muteVideo() ? 0.0 : 1.0), 10000);
             });
         }
@@ -134,11 +155,10 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
 
     protected onSystemInterrupt(interruptStarted: boolean): void {
         if(this.isShowing() && this.isActive()) {
-
             if(interruptStarted) {
                 this._nativeBridge.Sdk.logInfo('Pausing Unity Ads video playback due to interrupt');
                 this._nativeBridge.VideoPlayer.pause();
-            } else {
+            } else if (!interruptStarted && this.isVideoReady() && !this.getContainer().isPaused()) {
                 this._nativeBridge.Sdk.logInfo('Continuing Unity Ads video playback after interrupt');
                 this._nativeBridge.VideoPlayer.play();
             }