+++ src/ts/EventHandlers/VideoEventHandlers.ts
@@ -8,18 +8,23 @@ import { Diagnostics } from 'Utilities/Diagnostics';
 import { DiagnosticError } from 'Errors/DiagnosticError';
 import { VideoAdUnit } from 'AdUnits/VideoAdUnit';
 import { TestEnvironment } from 'Utilities/TestEnvironment';
-import { ViewConfiguration } from 'AdUnits/Containers/AdUnitContainer';
+import { AdUnitContainer, ViewConfiguration } from 'AdUnits/Containers/AdUnitContainer';
 import { Configuration } from 'Models/Configuration';
 import { Platform } from 'Constants/Platform';
 import { VideoMetadata } from 'Constants/Android/VideoMetadata';
+import { ViewController } from 'AdUnits/Containers/ViewController';
 
 export class VideoEventHandlers {
 
     public static onVideoPrepared(nativeBridge: NativeBridge, adUnit: VideoAdUnit, duration: number): void {
-        if(adUnit.getVideo().getErrorStatus()) {
+        if(adUnit.getVideo().getErrorStatus() || !adUnit.isPrepareCalled()) {
             // there can be a small race condition window with prepare timeout and canceling video prepare
             return;
         }
+
+        adUnit.getContainer().addDiagnosticsEvent({type: 'onVideoPrepared'});
+        adUnit.setPrepareCalled(false);
+        adUnit.setVideoReady(true);
 
         if(duration > 40000) {
             const campaign = adUnit.getCampaign();
@@ -66,16 +71,21 @@ export class VideoEventHandlers {
         nativeBridge.VideoPlayer.setVolume(new Double(overlay && overlay.isMuted() ? 0.0 : 1.0)).then(() => {
             if(adUnit.getVideo().getPosition() > 0) {
                 nativeBridge.VideoPlayer.seekTo(adUnit.getVideo().getPosition()).then(() => {
-                    nativeBridge.VideoPlayer.play();
+                    if(!adUnit.getContainer().isPaused()) {
+                        nativeBridge.VideoPlayer.play();
+                    }
                 });
             } else {
-                nativeBridge.VideoPlayer.play();
+                if(!adUnit.getContainer().isPaused()) {
+                    nativeBridge.VideoPlayer.play();
+                }
             }
         });
     }
 
     public static onVideoProgress(nativeBridge: NativeBridge, sessionManager: SessionManager, adUnit: VideoAdUnit, position: number, configuration: Configuration): void {
         if(position > 0 && !adUnit.getVideo().hasStarted()) {
+            adUnit.getContainer().addDiagnosticsEvent({type: 'videoStarted'});
             adUnit.getVideo().setStarted(true);
 
             sessionManager.sendStart(adUnit);
@@ -123,6 +133,7 @@ export class VideoEventHandlers {
 
                 // if video player has been repeating the same video position for more than 5000 milliseconds, video player is stuck
                 if(repeats > repeatTreshold) {
+                    adUnit.getContainer().addDiagnosticsEvent({type: 'videoStuck'});
                     nativeBridge.Sdk.logError('Unity Ads video player stuck to ' + position + 'ms position');
                     this.handleVideoError(nativeBridge, adUnit);
 
@@ -136,6 +147,11 @@ export class VideoEventHandlers {
                         cacheMode: configuration.getCacheMode(),
                         lowMemory: adUnit.isLowMemory()
                     };
+
+                    if(nativeBridge.getPlatform() === Platform.IOS && adUnit.getContainer() instanceof ViewController) {
+                        const container = <ViewController>adUnit.getContainer();
+                        error.events = container.getDiagnosticsEvents();
+                    }
 
                     const fileId = adUnit.getVideo().getFileId();
 
@@ -239,10 +255,12 @@ export class VideoEventHandlers {
     }
 
     public static onVideoPlay(nativeBridge: NativeBridge, adUnit: VideoAdUnit): void {
+        adUnit.getContainer().addDiagnosticsEvent({type: 'onVideoPlay'});
         nativeBridge.VideoPlayer.setProgressEventInterval(adUnit.getProgressInterval());
     }
 
     public static onVideoCompleted(sessionManager: SessionManager, adUnit: VideoAdUnit): void {
+        adUnit.getContainer().addDiagnosticsEvent({type: 'onVideoCompleted'});
         adUnit.setActive(false);
         adUnit.setFinishState(FinishState.COMPLETED);
         sessionManager.sendView(adUnit);
@@ -251,6 +269,7 @@ export class VideoEventHandlers {
     }
 
     public static onAndroidGenericVideoError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, what: number, extra: number, url: string) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onAndroidGenericVideoError', what: what, extra: extra});
         nativeBridge.Sdk.logError('Unity Ads video player error ' + ' ' + what + ' ' + extra + ' ' + url);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -264,6 +283,7 @@ export class VideoEventHandlers {
     }
 
     public static onIosGenericVideoError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, url: string, description: string) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onIosGenericVideoError', description: description});
         nativeBridge.Sdk.logError('Unity Ads video player generic error '  + url + ' ' + description);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -276,6 +296,7 @@ export class VideoEventHandlers {
     }
 
     public static onVideoPrepareTimeout(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, url: string): void {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onVideoPrepareTimeout'});
         nativeBridge.Sdk.logError('Unity Ads video player prepare timeout '  + url);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -287,6 +308,7 @@ export class VideoEventHandlers {
     }
 
     public static onPrepareError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, url: string) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onPrepareError'});
         nativeBridge.Sdk.logError('Unity Ads video player prepare error '  + url);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -298,6 +320,7 @@ export class VideoEventHandlers {
     }
 
     public static onSeekToError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, url: string) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onSeekToError'});
         nativeBridge.Sdk.logError('Unity Ads video player seek to error '  + url);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -309,6 +332,7 @@ export class VideoEventHandlers {
     }
 
     public static onPauseError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit, url: string) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onPauseError'});
         nativeBridge.Sdk.logError('Unity Ads video player pause error '  + url);
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -320,6 +344,7 @@ export class VideoEventHandlers {
     }
 
     public static onIllegalStateError(nativeBridge: NativeBridge, videoAdUnit: VideoAdUnit) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'onIllegalStateError'});
         nativeBridge.Sdk.logError('Unity Ads video player illegal state error');
 
         this.handleVideoError(nativeBridge, videoAdUnit);
@@ -329,7 +354,15 @@ export class VideoEventHandlers {
         });
     }
 
+    public static onIosVideoLikelyToKeepUp(nativeBridge: NativeBridge, adUnit: VideoAdUnit, container: AdUnitContainer, likelyToKeepUp: boolean): void {
+        adUnit.getContainer().addDiagnosticsEvent({type: 'onIosVideoLikelyToKeepUp', likelyToKeepUp: likelyToKeepUp});
+        if(!container.isPaused() && adUnit.getVideo().hasStarted() && likelyToKeepUp) {
+            nativeBridge.VideoPlayer.play();
+        }
+    }
+
     protected static afterVideoCompleted(adUnit: VideoAdUnit) {
+        adUnit.getContainer().addDiagnosticsEvent({type: 'afterVideoCompleted'});
         adUnit.getContainer().reconfigure(ViewConfiguration.ENDSCREEN);
 
         const overlay = adUnit.getOverlay();
@@ -340,6 +373,7 @@ export class VideoEventHandlers {
     }
 
     protected static updateViewsOnVideoError(videoAdUnit: VideoAdUnit) {
+        videoAdUnit.getContainer().addDiagnosticsEvent({type: 'updateViewsOnVideoError'});
         videoAdUnit.getContainer().reconfigure(ViewConfiguration.ENDSCREEN);
     }