+++ src/ts/Views/MRAID.ts
@@ -34,9 +34,6 @@ export class MRAID extends View {
     private _loaded = false;
 
     private _messageListener: any;
-    private _resizeHandler: any;
-    private _resizeDelayer: any;
-    private _resizeTimeout: any;
 
     private _canClose = false;
     private _canSkip = false;
@@ -66,19 +63,6 @@ export class MRAID extends View {
 
         const iframe: any = this._iframe = <HTMLIFrameElement>this._container.querySelector('#mraid-iframe');
 
-        if(this._nativeBridge.getPlatform() === Platform.IOS) {
-            if(Math.abs(<number>window.orientation) === 90) {
-                iframe.width = screen.height;
-                iframe.height = screen.width;
-            } else {
-                iframe.width = screen.width;
-                iframe.height = screen.height;
-            }
-        } else {
-            iframe.height = window.innerHeight;
-            iframe.width = window.innerWidth;
-        }
-
         this.createMRAID().then(mraid => {
             iframe.srcdoc = mraid;
         });
@@ -90,7 +74,6 @@ export class MRAID extends View {
     public show(): void {
         super.show();
 
-        const iframe: any = this._iframe;
         const closeLength = 30;
 
         if(this._placement.allowSkip()) {
@@ -147,30 +130,6 @@ export class MRAID extends View {
                 this.onLoaded.unsubscribe(observer);
             });
         }
-
-        this._resizeDelayer = (event: Event) => {
-            this._resizeTimeout = setTimeout(() => {
-                this._resizeHandler(event);
-            }, 200);
-        };
-
-        this._resizeHandler = (event: Event) => {
-            iframe.width = window.innerWidth;
-            iframe.height = window.innerHeight;
-            if(this._iframe.contentWindow) {
-                this._iframe.contentWindow.postMessage({
-                    type: 'resize',
-                    width: window.innerWidth,
-                    height: window.innerHeight
-                }, '*');
-            }
-        };
-
-        if(this._nativeBridge.getPlatform() === Platform.IOS) {
-            window.addEventListener('resize', this._resizeDelayer, false);
-        } else {
-            window.addEventListener('resize', this._resizeHandler, false);
-        }
     }
 
     public hide() {
@@ -182,15 +141,6 @@ export class MRAID extends View {
             window.removeEventListener('message', this._messageListener, false);
             this._messageListener = undefined;
         }
-        if(this._resizeHandler) {
-            window.removeEventListener('resize', this._resizeHandler, false);
-            this._resizeHandler = undefined;
-        }
-        if(this._resizeDelayer) {
-            window.removeEventListener('resize', this._resizeDelayer, false);
-            clearTimeout(this._resizeTimeout);
-            this._resizeHandler = undefined;
-        }
         super.hide();
     }
 
@@ -201,11 +151,30 @@ export class MRAID extends View {
                 if(markup) {
                     mraid = mraid.replace('{UNITY_DYNAMIC_MARKUP}', markup);
                 }
+                mraid = this.replaceMraidSources(mraid);
 
-                return MRAIDContainer.replace('<body></body>', '<body>' + mraid.replace('<script src="mraid.js"></script>', '') + '</body>');
+                return MRAIDContainer.replace('<body></body>', '<body>' + mraid + '</body>');
             }
             throw new WebViewError('Unable to fetch MRAID');
         });
+    }
+
+    public setViewableState(viewable: boolean) {
+        if(this._loaded) {
+            this._iframe.contentWindow.postMessage({
+                type: 'viewable',
+                value: viewable
+            }, '*');
+        }
+    }
+
+    private replaceMraidSources(mraid: string): string {
+        const dom = new DOMParser().parseFromString(mraid, "text/html");
+        const src = dom.documentElement.querySelector('script[src^="mraid.js"]');
+        if (src && src.parentNode) {
+            src.parentNode.removeChild(src);
+        }
+        return dom.documentElement.outerHTML;
     }
 
     private updateProgressCircle(container: HTMLElement, value: number) {
@@ -303,5 +272,4 @@ export class MRAID extends View {
             return Promise.resolve(this._campaign.getResource());
         }
     }
-
 }